leetcode/sort_and_searching_easy.py
from typing import List


def merge(nums1: List[int], m: int, nums2: List[int], n: int) -> None:
    p1 = m - 1
    p2 = n - 1
    tail = m + n - 1
    while p1 >= 0 or p2 >= 0:
        if p1 == -1:
            # 把nums 2 剩下的部分處理完成
            nums1[tail] = nums2[p2]
            p2 -= 1
        elif p2 == -1:
            # 把 num 1 剩下的部分處理完成
            nums1[tail] = nums1[p1]
            p1 -= 1
        elif nums1[p1] > nums2[p2]:
            nums1[tail] = nums1[p1]
            p1 -= 1
        else:
            nums1[tail] = nums2[p2]
            p2 -= 1
        tail -= 1
